Move RSA private key, tests to own file
